src/main/java/br/com/meridius/junit5/petclinic/entity/Person.java
package br.com.meridius.junit5.petclinic.entity;

public class Person extends BaseEntity {
	
	/** UID **/
	private static final long serialVersionUID = -7670050321691100689L;

	public Person(Long id, String firstName, String lastName) {
		super(id);
		this.firstName = firstName;
		this.lastName = lastName;
	}

	private String firstName;
	
	private String lastName;

	public String getFirstName() {
		return firstName;
	}

	public void setFirstName(String firstName) {
		this.firstName = firstName;
	}

	public String getLastName() {
		return lastName;
	}

	public void setLastName(String lastName) {
		this.lastName = lastName;
	}

	@Override
	public String toString() {
		return "Person [firstName=" + firstName + ", lastName=" + lastName + "]";
	}
}
